safari-600.7-branch/LayoutTests/plugins/npruntime/delete-plugin-within-setProperty-expected.txt

CONSOLE MESSAGE: line 13: Uncaught ReferenceError: NPObject deleted
